requirements.txt
langchain_community
streamlit
ipykernel
python-dotenv
langchain